Time: 0 ms (100%), Space: 8.6 MB (21.2%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.cpp
@@ -11,6 +11,6 @@ public:
         if (revNo > INT_MAX || revNo < INT_MIN) {
             return 0;
         }
-        return revNo;
+        return (int)revNo;
     }
 };
